Rm Analyzer::Config & ::EngineRegistry
These are no longer used in real code.

Regrettably, `config.rb` *was* being depended upon by some specs that
use it as source to report issues on! I am currenly a bit miffed at the
person who had that bright idea.
