+++ lib/cc/analyzer.rb
@@ -4,14 +4,12 @@ module CC
   module Analyzer
     autoload :Bridge, "cc/analyzer/bridge"
     autoload :CompositeContainerListener, "cc/analyzer/composite_container_listener"
-    autoload :Config, "cc/analyzer/config"
     autoload :Container, "cc/analyzer/container"
     autoload :ContainerListener, "cc/analyzer/container_listener"
     autoload :Engine, "cc/analyzer/engine"
     autoload :EngineOutput, "cc/analyzer/engine_output"
     autoload :EngineOutputFilter, "cc/analyzer/engine_output_filter"
     autoload :EngineOutputOverrider, "cc/analyzer/engine_output_overrider"
-    autoload :EngineRegistry, "cc/analyzer/engine_registry"
     autoload :EnginesConfigBuilder, "cc/analyzer/engines_config_builder"
     autoload :EnginesRunner, "cc/analyzer/engines_runner"
     autoload :Filesystem, "cc/analyzer/filesystem"
